refactor: ContentTypeRadio 이름 변경 및 리팩토링
- ContentTypeRadio, ContentTypeGroup -> ContentsRadio, ContentsRadioGroup으로 변경 (타입과 이름혼동 때문에)
- contentType 타입 정의 (ContentsRadio의 props 중 contentType 이름을 content로 바꿈)
- ContentsGroup은 map을 사용해서 코드를 줄였음
