+++ src/@types/contentType.js
@@ -1,0 +1,4 @@
+import { CONTENTS } from '@/constants';
+import { oneOf } from 'prop-types';
+
+export const contentType = oneOf(CONTENTS);
