+++ src/@types/index.js
@@ -1,1 +1,2 @@
 export * from './emotionType';
+export * from './contentType';
